+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Namespace io.pedestal.http.jetty.websockets has been removed, and replaced with io.pedestal.http.websockets
 - Namespace io.pedestal.interceptor.error has been moved to a new library, io.pedestal/pedestal.error
 - Library pedestal.service no longer has a dependency on org.clojure/core.match; the new pedestal.error library has that dependency
+- Interceptors that attach an invalid :response value (not a map, for example) will now cause an exception to be thrown
+- Interceptors may now attach a partial :response map, containing just a :status key
 
 Other changes:
 - When using io.pedestal.http/dev-interceptors, uncaught exceptions are now formatted using
@@ -32,6 +34,7 @@ Other changes:
 - Use of many deprecated functions and macros now cause deprecation warnings to be printed to stderr
 - Metrics and tracing have been reimplemented from the ground up around [Open Telemetry](https://opentelemetry.io/)
 - Libraries pedestal.log and pedestal.error contain [clj-kondo](https://github.com/clj-kondo/clj-kondo) configuration files
+- New function `io.pedestal.http/respond-with` to streamline adding a :response to the interceptor context
 
 [Closed Issues](https://github.com/pedestal/pedestal/milestone/12?closed=1)